Changed type variables

+++ src-tauri/src/models/ai_response.rs
@@ -10,10 +10,10 @@ pub struct AIResponse {
   pub created_at: String,
   pub message: AIMessage,
   pub done: bool,
-  pub total_duration: i32,
-  pub load_duration: i32,
-  pub prompt_eval_count: i32,
-  pub prompt_eval_duration: i32,
-  pub eval_count: i32,
-  pub eval_duration: i32
+  pub total_duration: i64,
+  pub load_duration: i64,
+  pub prompt_eval_count: i64,
+  pub prompt_eval_duration: i64,
+  pub eval_count: i64,
+  pub eval_duration: i64
 }
